added a new abstract class Schedule. all the schedules except CourseSectionSchedule extends Schedule
