+++ src/main/java/com/school_management/support_entities/schedule/StudentSchedule.java
@@ -1,56 +1,26 @@
 package com.school_management.support_entities.schedule;
 
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
-import java.util.Objects;
-
-public class StudentSchedule {
-    private List<CourseSectionSchedule> weeklySchedules;
+public class StudentSchedule extends Schedule{
 
     
     public StudentSchedule() {
-        this.weeklySchedules = new ArrayList<>();
+        super();
     }
 
-    public List<CourseSectionSchedule> getWeeklySchedules() {
-        return Collections.unmodifiableList(weeklySchedules);
-    }
-
-    public void addCourseSectionSchedule(CourseSectionSchedule schedule) {
-        for(CourseSectionSchedule cs: weeklySchedules) {
-            if(cs.hasConflict(schedule)) {
-                throw new IllegalArgumentException("Schedule conflicts with " + cs.toString());
-            } 
-        }
-        weeklySchedules.add(schedule);
-    }
-
-    public boolean removeCourseSectionSchedule(CourseSectionSchedule schedule) {
-        return weeklySchedules.remove(schedule);
-    }
 
     @Override
     public boolean equals(Object o) {
-        if (o == this)
-            return true;
-        if (!(o instanceof StudentSchedule)) {
-            return false;
-        }
-        StudentSchedule studentSchedule = (StudentSchedule) o;
-        return Objects.equals(weeklySchedules, studentSchedule.weeklySchedules);
+        return super.equals(o);
     }
 
     @Override
     public int hashCode() {
-        return Objects.hash(weeklySchedules);
+        return super.hashCode();
     }
 
     @Override
     public String toString() {
-        return "{" +
-            " weeklySchedules='" + getWeeklySchedules() + "'" +
-            "}";
+        return super.toString();
     }
